Change the default DllMain implementation
While refactoring the relationship between Metsrv and the extensions,
I removed the "reference" to Metsrv from all the extension projects. An
interesting error appeared that I thought was very strong. The error
said that the "Init()" function was missing, which is odd because the
_only_ DLL that uses RDI that requires the Init() function is Metsrv.
None of the extensions (or any other RDI DLLs in the exploits) should
have this at all.

This code change modifies the default behaviour such that all RDI DLLs
get a simple DllMain with standard behaviour. Given that Metsrv is the
exception, not the rule, this definitely makes the most sense. The
functionality required by Metsrv now lives only in Metsrv.